requirements.txt
flask==3.1.1
mysql-connector-python==0.3.1
ossapi==5.2.1
python-dotenv==1.1.0
requests==2.32.3